requirements.txt
absl-py==0.7.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
audioread==2.1.8
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.13.2
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
httplib2==0.9.2
idna==2.6
imageio==2.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.6
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
librosa==0.7.1
llvmlite==0.30.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.4
networkx==2.3
notebook==5.7.8
numba==0.46.0
numpy==1.16.4
oauth==1.0.1
olefile==0.45.1
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.7.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.0
pyRFC3339==1.0
pyrsistent==0.15.2
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.0.1
qtconsole==4.5.1
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
resampy==0.2.2
rope==0.14.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.11.0
sounddevice==0.3.14
SoundFile==0.10.2
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tflearn==0.3.2
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.6.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.1
xkit==0.0.0
zope.interface==4.3.2
